rythm-price-svc/requirements.txt
v20
python-dotenv
boto3
botostubs
kafka-python